+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- This fixes a small issue with the regex used to extract the time for e.g. break tags.
 ### Added 
 - Added attributes for say-as, prosody and emphasis tags and functions for parsing those
 - Added basic parsing of elements with no extracting of attributes for standard elements